{crucible,crux}-mir: Represent (), !, and FnDefs with MirAggregateType
Previously, these were represented with `UnitType`. This has the unfortunate
consequence that different zero-sized types were represented with different
Crucible type representations, as empty arrays are represented with
`MirAggregateType` instead. This poses an issue for some code in the internals
of `array::from_fn::<0>()`, which uses a `union` type that tries to reinterpret
a field of type `()` (represented with `UnitType`) as a length-0 array
(represented with `MirAggregateType`), and these Crucible type representations
are not compatible.

This changes the Crucible type representation of `()` (as well as `!` and
`FnDef`s, two other forms of zero-sized types) to instead use
`MirAggregateType`. This means that most zero-sized types now use
`MirAggregateType`, with the notable exception of empty structs (see #1499).
Still, things are much more uniform than they were before, which is a nice
improvement.

One slight annoyance brought about by this change is that several custom
overrides which previously returned `()` implicitly now must explicitly return
a `MirAggregate`. To make this slightly easier, I have introduced helper
functions `mirAggregate_zst{,IO,Sim}`, which can be used to construct
zero-sized `MirAggregate` values in such situations.

Fixes #1674.