crucible-llvm: Extremely limited support for poison constants
LLVM features a `poison` constant, which can have any type and theoretically
appear anywhere in an LLVM program. While Clang has historically avoided
generating `poison` constants in well-defined programs, more recent versions of
Clang have become more aggressive in inserting `poison` constants as a result
of optimization passes. For instance, the macOS Homebrew version of Clang 20
optimizes the `T1177.c` in such a way that bitcode of the following form is
generated:

```llvm
%50 = insertelement <8 x i32> <i32 poison, i32 0, i32 0, i32 0, i32 0, i32 0, i32 0, i32 0>, i32 %48, i64 0
```

Note the `poison` value in the first vector lane, which is then immediately
overwritten by the `insertelement` instruction. Prior to this patch,
`crucible-llvm` had absolutely no support for `poison` constants, which meant
that the bitcode above would crash `crucible-llvm`. This is an unfortunate
state of affairs, but the reality of the situation is that getting all the
details of tracking `poison` correctly is quite tricky. (See #366.)
Nevertheless, it feels like we could surely do at least slightly better here.

This patch is a small step in the direction of improving support for `poison`
constants. Like with `undef` values, we now add a dedicated constructs for
`poison` values in the `LLVMVal`, `LLVMExpr`, and `LLVMConst` data types, which
allow us to reason about `poison` constants at a very rudimentary level. We
also add `LLVM_{BV,Float}Poison` extension expressions that allow us to put
`poison` values in the Crucible simulator. For now, attempting to actually
simulate one of these `LLVM_{BV,Float}Poison` values will result in an error,
as we only support `poison` constants that are immediately overwritten (like in
the LLVM code snippet above). Accurately tracking `poison` values throughout
simulation is left as future work.