Remove crux-llvm's T1177.c test case in favor of crucible-llvm-cli tests
`crux-llvm`'s `T1177.c` test case was introduced in #1201 in order to test the
`llvm.vector.reduce.*` family of LLVM intrinsics. Unfortunately, this code also
tends to result in a lot of `poison` values being introduced as a result of
optimizations that recent versions of Clang use (e.g., macOS Homebrew's version
of LLVM 20). Because `crucible-llvm` cannot robustly handle `poison` values
at the moment (see #366), this poses an obstacle to testing.

Luckily, we can just as easily test the relevant LLVM intrinsics by writing
`crucible-llvm-ci` test cases, which exercise the relevant functionality in
`crucible-llvm` without being subject to the whims of Clang's optimizations.