crucible-llvm: Remove the Eq LLVMConst instance
In the presence of `undef` and `poison` values, there is no sensible way to
always return a `True`-or-`False` answer to the question of whether two
`LLVMConst`s are equal or not. This makes the existence of an `Eq LLVMConst`
instance inherently dodgy.

Luckily, this instance is not used anywhere in the `crucible` repo, and I would
be very surprised if it were used elsewhere. Let's just remove the instance.