crucible-llvm: Remove the Eq LLVMConst instance
In the presence of `undef` and `poison` values, there is no sensible way to
always return a `True`-or-`False` answer to the question of whether two
`LLVMConst`s are equal or not. This makes the existence of an `Eq LLVMConst`
instance inherently dodgy.

This instance is used in exactly one place in order to check `LLVMConst`s for
syntactic equality in the `crucible-llvm` test suite. (I'd be very surprised if
it were used elsewhere outside of the repo.) As such, I've decided to remove
the instance and instead define a `llvmConstSyntacticEq` function for use in
the `crucible-llvm` test suite. (We could expose this function as part of the
`crucible-llvm` API, but it doesn't seem that generally useful, so I have opted
not to do this.)